Continue integrating question storage [in progress]
