+++ src/main/java/org/vstu/compprehension/models/businesslogic/storage/RdfStorage.java
@@ -50,7 +50,9 @@ import java.util.stream.Stream;
 
 
 /**
- * Use separate RdfStorage for each domain.
+ * Storage of templates & questions for a domain.
+ * Metadata is available from SPARQL endpoint,
+ * Actual questions data is accessed as files by FTP/HTTP/local filesystem.
  */
 @Log4j2
 public class RdfStorage extends AbstractRdfStorage {
@@ -68,7 +70,7 @@ public class RdfStorage extends AbstractRdfStorage {
     }
 
     /**
-     * Full path with name of directory under BASE_DB_PATH to store domain-specific database in
+     * Relative path under FUSEKI_ENDPOINT_BASE to store domain-specific database in
      */
     String sparql_endpoint = null;
 
@@ -142,179 +144,6 @@ public class RdfStorage extends AbstractRdfStorage {
         }
     }
 
-    /**
-     * Find questions in current questions bank
-     * @param qr QuestionRequest
-     * @param limit maximum questions to return (set 0 or less to disable; note: this may be slow)
-     * // (ignore randomSeed) param randomSeed influence randomized order (if selection from many candidates is required)
-     * @return questions found or empty list if the requirements cannot be satisfied
-     */
-    @Override
-    List<Question> searchQuestions(QuestionRequest qr, int limit) {
-        Model qG = getGraph(NS_questions.base());
-
-        String complexity_threshold = "" + qr.getComplexity();
-        String complexity_cmp_op = (qr.getComplexitySearchDirection().getValue() < 0)? "<=" : ">=";
-
-        List<Double> SolutionLengthStat = questionSolutionLengthStat();  // [min, avg, max]
-        int duration = qr.getSolvingDuration();  // 1..10
-        int desiredSolutionSteps = (int) Math.round(SolutionLengthStat.get(1));  // avg
-        if (duration < 5) {
-            desiredSolutionSteps = (int) Math.round(
-                    SolutionLengthStat.get(0) +
-                            ((duration - 1) / 4.0) * (SolutionLengthStat.get(1) - SolutionLengthStat.get(0))
-            );  // min + weight * (avg - min)
-        }
-        else if (duration > 5) {
-            desiredSolutionSteps = (int) Math.round(
-                    SolutionLengthStat.get(1) +
-                            ((duration - 5) / 5.0) * (SolutionLengthStat.get(2) - SolutionLengthStat.get(1))
-            );  // avg + weight * (max - avg)
-        }
-
-        StringBuilder query = new StringBuilder("select distinct ?name ?complexity ?solution_steps \n" +
-                //// "#(count(distinct ?law) as ?law_count) (count(distinct ?concept) as ?concept_count)\n" +
-                "(group_concat(distinct ?law;separator=\"; \") as ?laws)\n" +
-                "(group_concat(distinct ?concept;separator=\"; \") as ?concepts)\n" +
-                " {  GRAPH <" + NS_questions.base() + "> {\n" +
-                "\t?s a <" + NS_questions.get("Question") + "> .\n" +
-                "    ?s <" + NS_questions.get("name") + "> ?name.\n" +
-                "    ?s <" + NS_questions.get("integral_complexity") + "> ?complexity.\n" +
-                "    ?s <" + NS_questions.get("solution_steps") + "> ?solution_steps.\n" +
-                "    ?s <" + NS_questions.get("has_concept") + "> ?concept.\n" +
-                "    ?s <" + NS_questions.get("has_violation") + "> ?law.\n" +
-                "    filter(?complexity " + complexity_cmp_op + " " + complexity_threshold + ").\n" +
-                "    bind(abs(" + complexity_threshold + " - ?complexity) as ?compl_diff).\n" +
-                "    bind(abs(" + desiredSolutionSteps + " - ?solution_steps) as ?steps_diff).\n");
-        // add "denied" constraints
-            //  "    filter not exists { ?s <has_concept> \"denied_concept\" }\n" +
-            //  "    filter not exists { ?s <http://vstu.ru/poas/questions/has_law> \"denied_law\" }\n" +
-        for (Concept concept : qr.getDeniedConcepts()) {
-            String denied_concept = concept.getName();
-            query.append("    filter not exists { ?s <")
-                    .append(NS_questions.get("has_concept"))
-                    .append("> \"")
-                    .append(denied_concept).append("\" }\n");
-        }
-        for (Law law : qr.getDeniedLaws()) {
-            String denied_law = law.getName();
-            query.append("    filter not exists { ?s <")
-                    .append(NS_questions.get("has_violation"))
-                    .append("> \"")
-                    .append(denied_law).append("\" }\n");
-        }
-        if (qr.getLawsSearchDirection().getValue() > 0) {
-            //  "#    filter exists { ?s <http://vstu.ru/poas/questions/has_law> \"target_law\" }\n";
-            for (Law law : qr.getTargetLaws()) {
-                String target_law = law.getName();
-                query.append("    filter exists { ?s <")
-                        .append(NS_questions.get("has_violation"))
-                        .append("> \"")
-                        .append(target_law).append("\" }\n");
-            }
-        }
-//                "#    filter exists { ?s <http://vstu.ru/poas/questions/has_concept> \"target_concept\" }\n" +
-        // query footer
-        query.append("  }}\n" +
-                "group by ?name ?complexity ?solution_steps ?compl_diff\n" +
-                "order by ?compl_diff ?steps_diff");
-        if (limit > 0)
-            query.append("\n  limit " + limit * 2);
-
-        List<Question> questions = new ArrayList<>();
-        try ( RDFConnection conn = RDFConnection.connect(dataset) ) {
-
-            List<Question> finalQuestions = questions;  // copy reference as the lambda syntax requires
-            conn.querySelect(query.toString(), querySolution -> {
-                QuestionEntity qe = new QuestionEntity();
-                qe.setQuestionName( querySolution.get("name").asLiteral().getString() );
-                // todo: pass the kind of question here
-                Question q = new Ordering(qe);
-                q.getConcepts().addAll( List.of(querySolution.get("concepts").asLiteral().getString().split("; ")) );
-                q.getNegativeLaws().addAll( List.of(querySolution.get("laws").asLiteral().getString().split("; ")) );
-
-                if (limit > 0) {  // don't calc score if no filtering is required
-                    int score = 0;  // how the question suits the request
-
-                    // inspect laws
-                    if (qr.getLawsSearchDirection().getValue() < 0) {
-                        int i = 0;
-                        int size = qr.getTargetLaws().size();
-                        score += size * size / 2;  // max possible score for existing
-                        // set penalty for each missing law (earlier ones cost more)
-                        for (Law target : qr.getTargetLaws()) {
-                            i += 1;
-                            if (!q.getNegativeLaws().contains(target.getName()))
-                                score -= (size - i);
-                        }
-                    }
-                    // set penalty for each extra law that is not in allowed laws
-                    ArrayList<String> extraLaws = new ArrayList<>(q.getNegativeLaws());
-                    extraLaws.removeAll(qr.getTargetLaws().stream().map(Law::getName).collect(Collectors.toList()));
-                    if (!extraLaws.isEmpty()) {
-                        // found out how many laws are "not allowed"
-                        extraLaws.removeAll(qr.getAllowedLaws().stream().map(Law::getName).collect(Collectors.toList()));
-                        score -= extraLaws.size();  // -1 for each extra law
-                    }
-
-                    // inspect concepts
-                    int i = 0;
-                    int size = qr.getTargetConcepts().size();
-                    score += size * size / 2;  // max possible score for existing
-                    // set penalty for each missing law (earlier ones cost more)
-                    for (Concept target : qr.getTargetConcepts()) {
-                        i += 1;
-                        if (!q.getConcepts().contains(target.getName()))
-                            score -= (size - i);
-                    }
-
-                    // set penalty for each extra Concepts that is not in allowed laws
-                    ArrayList<String> extraConcepts = new ArrayList<>(q.getConcepts());
-                    extraConcepts.removeAll(qr.getTargetConcepts().stream().map(Concept::getName).collect(Collectors.toList()));
-                    if (!extraConcepts.isEmpty()) {
-                        // found out how many Concepts are "not allowed"
-                        extraConcepts.removeAll(qr.getAllowedConcepts().stream().map(Concept::getName).collect(Collectors.toList()));
-                        score -= extraConcepts.size();  // -1 for each extra concept
-                    }
-
-                    qe.setId((long) score);  // (ab)use ID for setting score temporarily
-                }
-
-                // add some data about this question
-                qe.getStatementFacts().add(new BackendFactEntity("<this_question>", "complexity", "" + querySolution.get("complexity").asLiteral().getDouble()));
-                qe.getStatementFacts().add(new BackendFactEntity("<this_question>", "solution_steps", "" + querySolution.get("solution_steps").asLiteral().getDouble()));
-
-                finalQuestions.add(q);
-            });
-
-        } catch (JenaException exception) {
-            exception.printStackTrace();
-        }
-
-        if (limit > 0) {
-            // sort to ensure best score first
-            questions.sort((o1, o2) -> -(int) (o1.getQuestionData().getId() - o2.getQuestionData().getId()));
-            // retain requested number of questions
-            questions = questions.subList(0, limit);
-        }
-
-        // insert to the questions some data (all we can here)
-        for (Question q : questions) {
-            QuestionEntity qe = q.getQuestionData();
-            qe.setDomainEntity(domain.getEntity());
-            qe.setExerciseAttempt(qr.getExerciseAttempt());
-            //// qe.setOptions(domain.?);
-
-            qe.setStatementFacts( modelToFacts(
-                    getQuestionModel(q.getQuestionName()),
-                    // TODO: adapt base URI for each domain
-                    NS_code.base()
-                    ));
-        }
-
-        return questions;
-    }
-
 
     /*
             REMOTE DATABASE COMMUNICATION
@@ -326,25 +155,6 @@ public class RdfStorage extends AbstractRdfStorage {
         RDFConnectionRemoteBuilder cb = RDFConnectionFuseki.create()
                 .destination(sparql_endpoint);
         return cb.build();
-    }
-
-    /** Download, cache and return a graph */
-    @Override
-    @Nullable
-    Model getGraph(String name) {
-        if (fetchGraph(name)) {
-            return getLocalGraphByUri(name);
-        }
-        return null;
-    }
-
-    /** Cache and send a graph to remote DB */
-    @Override
-    boolean sendGraph(String gUri, Model m) {
-        if (setLocalGraph(gUri, m)) {
-            return uploadGraph(gUri);
-        }
-        return false;
     }
 
     /** Download and cache a graph
@@ -486,42 +296,7 @@ public class RdfStorage extends AbstractRdfStorage {
     }
 
 
-    public boolean localGraphExists(String gUri) {
-        return dataset.containsNamedModel(gUri);
-    }
-
-    public Model getLocalGraphByUri(String gUri) {
-        if (dataset != null) {
-            if (localGraphExists(gUri)) {
-                return dataset.getNamedModel(gUri);
-            } else log.warn(String.format("Graph not found - name: '%s'", gUri));
-        } else {
-            log.warn("Dataset was not initialized");
-        }
-        return null;
-    }
-
-    /**
-     * запись/обновление подграфа триплетов в кэше
-     * @param gUri graph URI
-     * @param model model to set
-     * @return true on success
-     */
-    public boolean setLocalGraph(String gUri, Model model) {
-        if (dataset != null) {
-            if (dataset.containsNamedModel(gUri))
-                dataset.replaceNamedModel(gUri, model);
-            else
-                dataset.addNamedModel(gUri, model);
-
-        } else {
-            log.error("Dataset was not initialized");
-            throw new RuntimeException("Dataset was not initialized");
-        }
-        return true;
-    }
-
-//    /**
+    //    /**
 //     * получение подграфа триплетов, как он есть в хранилище
 //     * @param localName graph name
 //     * @return model of triples
@@ -548,47 +323,6 @@ public class RdfStorage extends AbstractRdfStorage {
     //}
 
 
-    /**
-     * получение подграфа триплетов, хранящего базовую схему домена (необходимую для работы ризонера)
-     * @return model of triples
-     */
-    @Override
-    public Model getSchema() {
-        String uri = NS_graphs.get(GraphRole.SCHEMA.ns().base());
-        if (!dataset.containsNamedModel(uri)) {
-            setLocalGraph(uri, getDomainSchemaForSolving());
-        }
-        return getLocalGraphByUri(uri);
-    }
-
-
-
-    /**
-     * Get solved domain schema (for reasoning purposes)
-     * @return model both schema and solved schema (the most of what exists - may be empty)
-     */
-    @Override
-    public Model getFullSchema() {
-        Model m = ModelFactory.createDefaultModel();
-        Model m2 = getSchema();
-        if (m2 != null)
-            m.add(m2);
-
-        String uri = NS_graphs.get(GraphRole.SCHEMA_SOLVED.ns().base());
-        if (!dataset.containsNamedModel(uri) && !m.isEmpty()) {
-            Model inferred = runReasoning(m, getDomainRulesForSolvingAtLevel(GraphRole.SCHEMA), true);
-            if (inferred.isEmpty()) {
-                // add anything to avoid re-calculation
-                inferred.add(m.listStatements().nextStatement());
-            }
-            setLocalGraph(uri, inferred);
-        }
-        m2 = getLocalGraphByUri(uri);
-        if (m2 != null)
-            m.add(m2);
-
-        return m;
-    }
 
     /* *
      * получить модель, объединённую из всех подграфов для вопроса
